catalog-agent/requirements.txt
langgraph==0.2.60
langchain-google-genai==2.0.0
pydantic==2.10.0
motor==3.6.0
python-dotenv==1.0.1
temporalio==1.7.0
httpx==0.27.0
tqdm==4.66.0
